+++ CHANGELOG.md
@@ -479,6 +479,8 @@ the magnitude processing: ```EpicenterOutOfRegions```.
   * Map layer drawing properties may be additionally defined in a "map.cfg" file
     located in the data set folder and subfolder, e.g. @~/.seiscomp3/fep/map.cfg@,
     @~/.seiscomp3/bna/map.cfg@, @~/.seiscomp3/bna/category/map.cfg@
+  * Added support for event summary to listen to alert comments and adapt size and color
+    of time ago label accordingly.
 
 * scmv